feat: add layers and functions for privado-core
This commit involves modifying existing scan and adds new functions to interact with privado-core based options. It also removes the previous cmd commands which are now redacted and not relevant for the upcoming release cycle. We will continue to make modifications to the interaction layer to migrate the image for new use-cases.