feat: add container output processor; setup interrupt
For container output processor to work without missing output, we tweaked how output was printed to console (from container stdout, stderr to os.stdout). We now simply print and the outputs instead of copying them in a parallel goroutine, as we noticed some output to be missing and realized a much complex nature of readers, this might also be contributed by the fact that we were processing each output line serially, however the independent goroutine might not wait.

To conclude, we now use parallel processsing of output lines and use print for container output per line. During implementation, we noticed a memory leak where containers did not terminate on Ctrl+C and thereby also enabled interrupt fn during scan.